test(): add a few to Controllers to show how to use the framework
TODO: move this directory to the root of the project and rename it in 'examples'. The Makefile
will have a new rule to make examples: each example will be compiled as a single program (so
they will all need a main function). The framework will be staticaly linked to the examples (just
compile with all *.o files compiled from the framework).